third_party/blink/web_tests/editing/execCommand/reset-values-after-navigation.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script>
    window.jsTestIsAsync = true;
    window.onmessage = function (e) {
        queryCommandValues = JSON.parse(e.data);
        shouldBeEqualToString('queryCommandValues["DefaultParagraphSeparator"]', "div");
        shouldBeEqualToString('queryCommandValues["StyleWithCSS"]', "");
        finishJSTest();
    }
</script>
</head>
<body>
<iframe style="display: none" src="resources/reset-default-values-helper-1.html"></iframe>
</body>
</html>
